Update documentation, especially add comments to public API
This change adds documentation to all public methods. This makes it
easier to reason about the code and increases understanding.
The way the documentation is provided should allow IDEs and editors to
pick it up and provide it interactiely.
Furthermore this change split some external types up to make it easier
to distinguish between them in the implementation. This does _not_ change
the API.