Allow comparison of tuples by their first item
This change fixes the comparison of tuples and arrays if the first item
should be compared.
Prior to this change, the comparison was only possible for elements of
index 1 or higher. The reason for this mismatch was an insufficient
check of the comparator is capable of extracting valid values from the
compared objects. The new implementation explicitly excludes zero from
this check, to make sure that the validator works as expected.

Note that this problem might happen with an empty string as well. But to
be fair, I don't expect someone to hard-code an empty string on an
object as a key.

Closes: #1