Replace gts by @antfu/eslint-config for code-formatting
In the past I constantly failed to submitted a properly formatted
code-base. Mostly, because I used gts as a formatting and linting
solution which doesn't support me directly in my IDE. Instead I had to
run `npm run fix` all the time manually. Furthermore gts was using an
dated eslint version which was also annoying.

The new setup is an ease to use, and I can easily adopt my IDE to
autoformat the code correctly. It also fulfills the other requirement,
to be usable with an up-to-date version of eslint. On top of that I was
able to get rid of prettier this way which is a step back for most
code-bases in my opinion.

Finally the new setup lead to a formatting which I like more than the
old one and is still very concise and instead even improves the
readability.