Do not send slack-notification when npm-update failed
We run a check for npm-updates daily but have so few dependencies, that
we can expect to see way more "update failed" messages than "updated"
messages. This lead to the unfortunate situation that we miss our
important event, the update, in all these spam-notifications.

This change fixes the described problem by only notifying about
successful updates.