Maintain version in a branch of this repository
The version was previously maintained in a dedicated repository. I think
that is a bit pointless, since it makes it way too easy to damage the
integrity of this project.

So, this change simply uses an orphan branch "version" of this
repository to maintain the version. That way the version is handled way
more concise.